recursos/leccion2/mysql/sistema/requirements.txt
# Install the packages from PyPi repository.
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography==44.0.1
